Добавил методы toString(), equals(), hashCode() в классы Author и Book Переписал метод printBook в главном классе.
